+++ lib/src/format.rs
@@ -3,8 +3,8 @@
 /// Deprecated compatibility layer.
 ///
 /// ```
-/// #// If this test fails, update the deprecation notice too
-/// use crate::config::ReplayBufferKind;
+/// # // If this test fails, update the deprecation notice too
+/// use libtechmino_replay::config::ReplayBufferKind;
 /// ```
 #[deprecated(
     since = "0.2.0",
@@ -15,8 +15,8 @@ pub use crate::config::ReplayBufferKind;
 /// Deprecated compatibility layer.
 ///
 /// ```
-/// #// If this test fails, update the deprecation notice too
-/// use crate::replay::SerializedReplay;
+/// # // If this test fails, update the deprecation notice too
+/// use libtechmino_replay::replay::SerializedReplay;
 /// ```
 #[deprecated(
     since = "0.2.0",